asdf

+++ math.h
@@ -8,7 +8,7 @@
 template <class T=double> 
 T sum(std::vector<T> w, connection<> a) {
     T result = 0;
-    for (int i = 0; i < w.size(); i++) {
+    for (size_t i = 0; i < w.size(); i++) {
         result += *a[i] * w[i];
     }
     return result;
